Move `View_Only_Pointer` from `Core\Util` to `Core\Dashboard_Sharing` namespace.
